started repl completion
