+++ src/Makefile
@@ -2,7 +2,7 @@ CC?=gcc
 VERSION := $(shell cat ../VERSION)
 
 CFLAGS  += -I. -I../lib/lua_sandbox/include -O3 -Wall -Wextra
-SOURCES := jutils.o lua_functions.o zenroom.o bitop.o lua_cjson.o strbuf.o fpconv.o repl.o linenoise.o cheatsheet.o
+SOURCES := jutils.o lua_functions.o zenroom.o bitop.o lua_cjson.o strbuf.o fpconv.o repl.o replcompl.o linenoise.o cheatsheet.o
 LDADD   := ../build/lua_sandbox/src/libluasandbox.a
 LDADD   += ../build/lua_sandbox/src/util/libluasandboxutil.a
 LDADD   += ../build/luazen/libluazen.a
